<refactor> переструктуризация
