+++ sast-agent/pom.xml
@@ -9,7 +9,7 @@
 	</parent>
 
 	<artifactId>sast-agent</artifactId>
-	<name>sast-sec</name>
+	<name>sast-agent</name>
 
 	<packaging>jar</packaging>
 
@@ -17,16 +17,7 @@
 		<java.version>21</java.version>
 	</properties>
 	<dependencies>
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-security</artifactId>
-		</dependency>
 
-		<dependency>
-			<groupId>org.springframework.security</groupId>
-			<artifactId>spring-security-test</artifactId>
-			<scope>test</scope>
-		</dependency>
 	</dependencies>
 
 	<build>
