+++ sast-broker/Dockerfile
@@ -5,4 +5,4 @@ VOLUME /tmp
 COPY target/sast-client-0.0.1-SNAPSHOT.jar sast-client-0.0.1-SNAPSHOT.jar
 ENTRYPOINT ["java", "-jar", "/sast-client-0.0.1-SNAPSHOT.jar"]
 
-EXPOSE 8080
+EXPOSE 7070
