+++ sast-broker/src/main/java/ru/SberTex/SastBroker/Main.java
@@ -1,4 +1,4 @@
-package ru.SberTex.SastClient;
+package ru.SberTex.SastBroker;
 
 public class Main {
     public static void main(String[] args) {
