+++ sast-manager/Dockerfile
@@ -5,4 +5,4 @@ VOLUME /tmp
 COPY target/sast-service-0.0.1-SNAPSHOT.jar sast-service-0.0.1-SNAPSHOT.jar
 ENTRYPOINT ["java", "-jar", "/sast-service-0.0.1-SNAPSHOT.jar"]
 
-EXPOSE 9090
+EXPOSE 8080
