+++ sast-manager/pom.xml
@@ -9,7 +9,7 @@
 	</parent>
 
 	<artifactId>sast-manager</artifactId>
-	<name>sast-server</name>
+	<name>sast-manager</name>
 
 	<packaging>jar</packaging>
 
@@ -20,6 +20,16 @@
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-data-jpa</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-security</artifactId>
+		</dependency>
+
+		<dependency>
+			<groupId>org.springframework.security</groupId>
+			<artifactId>spring-security-test</artifactId>
+			<scope>test</scope>
 		</dependency>
 	</dependencies>
