+++ sast-manager/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-server.port=9090
+server.port=8080
 spring.jpa.hibernate.ddl-auto=none
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.properties.hibernate.format_sql=true
